feat: fuel front page with no form, renaming imgs
