+++ src/components/Fuel/FuelForm.tsx
@@ -1,10 +1,16 @@
 import React from "react";
+import { useState } from "react";
 
 export default function FuelForm () {
+/* 
+    const [fuelData, setfuelData] = useState({
+
+                                    }) */
+
     return (
         <div>
             {/* Replace this placeholder route */}
-            <form action="/api/form" method="POST">
+            <form action="/api/form">
                 
                 {/* Add regex to validate answers once I know what the questions are. */}
                 <label htmlFor="api_name">Which fuel do you use?</label>
@@ -15,7 +21,7 @@ export default function FuelForm () {
                     <option value="ker">Kerosene</option>
                 </select>
 
-                {/* COnditionally render the options according to: https://www.notion.so/Carbon-Interface-Fuel-Sources-0166b59ec1514984895cc7dd35836392 */}
+                {/* Conditionally render the options according to: https://www.notion.so/Carbon-Interface-Fuel-Sources-0166b59ec1514984895cc7dd35836392 */}
                 <label htmlFor="unit">Which measuring unit do you use?</label>
                 <select name="unit" id="unit">
                     <option value="btu">BTU</option>
