--- src/components/dummyComponent.tsx
@@ -1,11 +1,0 @@
-// JUST TO CREATE THE FOLDER. THIS IS A GARBAGE FILLER COMPONENT
-
-import React from 'react'
-
-const dummyComponent = () => {
-  return (
-    <div>dummyComponent</div>
-  )
-}
-
-export default dummyComponent
