+++ src/pages/fuel.tsx
@@ -1,21 +1,29 @@
 import Link from 'next/link'
 import Head from 'next/head'
-import FuelForm from '../components/FuelUsed/FuelForm'
+import Image from 'next/image'
+import plantingTreesImage from '../assets/images/planting-trees-1.png'
+import FuelForm from '../components/Fuel/FuelForm'
 
-export default function FuelConsumption() {
+
+export default function FuelPage() {
   return (
     <div>
       <Head>
         <title>Fuel Consumption</title>
         <meta name="viewport" content="initial-scale=1.0, width=device-width" />
       </Head>
-      <picture>
-        <img src="../assets/images/watering-plants.png" alt="" />
-      </picture>
-      
-      <h2>How much carbon your fuel intake has emitted so far?</h2>
-      <FuelForm />
-  
+
+      <div className='flex justify-center'>
+        <Image src={plantingTreesImage} alt="Four people planting trees"/>
+      </div>
+
+      <h2 className='text-4xl flex justify-center gap-4 w-100'>How much carbon your fuel intake has emitted so far?</h2>
+
+      <div  className="container mx-auto flex justify-center">
+        <FuelForm />
+      </div>
+     
+
     </div>
   )
 }
