+++ src/pages/travel.tsx
@@ -3,7 +3,7 @@ import Head from 'next/head';
 import { Accordion, Tabs } from 'flowbite-react';
 import { FaPlane, FaCarAlt } from 'react-icons/fa';
 import FlightQuestionaire from '../components/Flight/FlightQuestionaire';
-import wateringPlantsImage from '../assets/images/watering-plants-removebg-preview 1.png';
+import wateringPlantsImage from '../assets/images/watering-plants.png';
 import Image from 'next/image';
 
 export default function TravelPage() {
@@ -27,7 +27,6 @@ export default function TravelPage() {
             <FlightQuestionaire />
           </Tabs.Item>
           <Tabs.Item title="Driving" icon={FaCarAlt}>
-            <p>Your Driving parent component here</p>
             <FlightQuestionaire />
           </Tabs.Item>
         </Tabs.Group>
